+++ CHANGES.md
@@ -9,6 +9,7 @@ Apollo Java 2.5.0
 * [Feature Provide a new open APl to return the organization list](https://github.com/apolloconfig/apollo-java/pull/102)
 * [Feature Added a new feature to get instance count by namespace.](https://github.com/apolloconfig/apollo-java/pull/103)
 * [Feature Support retry in open api client.](https://github.com/apolloconfig/apollo-java/pull/105)
+* [Support Spring Boot 4.0 bootstrap context package relocation for apollo-client-config-data](https://github.com/apolloconfig/apollo-java/pull/115)
 
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo-java/milestone/5?closed=1)